CpG network: edges are based on CpG island affiliation

For every present CGI, function "meNet_singleCGI" is called and the resulting network is a union of the single-CGI networks. (Optional) weights are chromosomal distances.

meNet_CGI(
  cg_list = NULL,
  cgi_list = NULL,
  link_method = "twoLyr_clust",
  weighted = TRUE,
  cor_matrix = NULL,
  data = NULL,
  cor_normalization_fun = max_normalization,
  dist_normalization_fun = neg_max_normalization,
  cor_threshold = 0.2,
  neg_cor_threshold = NULL,
  cor_stDev = NULL,
  cor_alpha = NULL,
  n_repetitions = 1000,
  alternative = "two_sided",
  infomap_call = "infomap",
  folder = "./meNet/",
  file_basename = "meNet_CGI_infomap",
  relaxation_rate = 0.15,
  cg_meta = data("CpG_anno450K", package = "meNet"),
  cg_meta_cols = list(cg_id = "IlmnID", cg_coord = "MAPINFO", island_name =
    "UCSC_CpG_Islands_Name", island_region = "Relation_to_UCSC_CpG_Island"),
  include_regions = character(0),
  expand_cg_list = FALSE,
  normalization_fun = NULL,
  save_all_files = FALSE,
  delete_files = FALSE
)

Arguments

cg_list
Vector of CpG sites.

cgi_list
Vector of CpG islands.

link_method
Method used to determine the edges within each CGI subnetwork. See details. Default value is `"twoLyr_clust"`.

weighted
Whether the resulting network is weighted. If `TRUE`, the weights are base pair distances between CpGs. Defaults to `TRUE`.

cor_matrix
Correlation matrix of CpG sites.

data
Data frame with CpGs in columns. Variables in rows are used to calculate `cor_matrix`.

cor_normalization_fun
Normalization function for the correlation layer of a CGI. Default method is `max_normalization` function.

dist_normalization_fun
Normalization function for the distance layer of a CGI. Default method is `neg_max_normalization` function.

cor_threshold
Correlation threshold. Defaults to `0.2`.

neg_cor_threshold
Negative correlation threshold. Defaults to `NULL`.

cor_stDev
Threshold for the standard deviation of correlation. Defaults to `NULL`.

cor_alpha
Significance level of the correlation permutation test. Defaults to `NULL`.

n_repetitions
Number of repetitions for resampling and/or for the correlation permutation test. Defaults to `1000`.

alternative
Alternative hypothesis for the correlation permutation test. Default value is `"two_sided"`.

infomap_call
Path to the `Infomap` on user's system. Default value is `"infomap"`.

folder
Folder in which the files will be saved. It is automatically created if not already present. Default value is `"./meNet/"`.

file_basename
Base name of the created files. Default value is `"meNet_CGI_infomap"`. If `save_all_files` is `TRUE`, its value changes for every CGI and becomes equal to name of the CGI.

relaxation_rate
`multilayer-relax-rate` parameter in `Infomap` call. Defaults to `0.15`.

cg_meta
Data frame which defines relationship between CpG sites and CpG islands. It has CpGs in the rows and their description in the columns. If CpG is part of an CGI, `cg_meta` gives the name of the island and the region of an island in which the CpG is found. Additionally, if base pair distances are being calculated, `cg_meta` also holds chromosomal coordinate of CpGs. By default, function uses `CpG_anno450K` data frame which contains Illumina Infinium HumanMethylation450 manifest file.

cg_meta_cols
Named list with `cg_meta` column names. The list must include: `cg_id` naming the column with unique CpG names, `island_name` naming the column with CGI names and `island_region` naming the column with the CGI region in which the CpG is located. If base pair distances are being calculated, it should also include `cg_coord` naming the column with chromosomal coordinates. Default value shouldn't be changed if `cg_meta` keeps it default value.

include_regions
Regions of CGIs which should be searched for CpGs besides the island itself. Can take any subset of values `"S_Shelf"`, `"S_Shore"`, `"N_Shore"`, `"N_Shelf"`. By default, only the CpGs inside the island are reported.

expand_cg_list
Whether, for a CGI containing a CpG from `cg_list`, all CpGs should automatically be added to the network. Default to `FALSE`.

normalization_fun
Normalization function applied on the weights of the resulting network, if the network is weighted.

save_all_files
Whether all created files should be saved. Defaults to `FALSE`.

delete_files
Should created files be automatically deleted from the user's system. Defaults to `FALSE`. If `save_all_files` is `TRUE` then its value is automatically changed to `FALSE.` Changing the parameter to `TRUE` should be done with caution since it will allow the function to delete files from user's system.

Value

Details

List of nodes can be given either as list of CpGs with parameter `cg_list` or as list of CpG islands with parameter `cgi_list`. In case both lists are given, union of nodes is used.

Passed to meNet_singleCGI `link_method`, `weighted`, `cor_normalization_fun`, `dist_normalization`, `cor_threshold`, `neg_cor_threshold`, `cor_stDev`, `cor_alpha`, `n_repetitions`, `alternative`, `infomap_call`, `folder`, `relaxation_rate`, `cg_meta`, `cg_meta_cols`, `include_regions`, `check_matrices`, `delete_files`