Beta value error plot with CpG coordinates

Description

plotCI_meCpG(
  beta_values,
  sample_groups,
  cg_list = NULL,
  chr = NULL,
  first_coord = NULL,
  last_coord = NULL,
  plot_cg = FALSE,
  plot_cgi = FALSE,
  plot_gene = FALSE,
  plot_exon = FALSE,
  transcript_types = c("NM", "NR"),
  col_groups = NULL,
  col_gene = "green",
  col_cgi = "magenta",
  title = "",
  x_label = "",
  y_label = "",
  text_cex = 1,
  beta_min = 0,
  beta_max = 1,
  error_size = 1,
  plot_legend = FALSE,
  plot_outside = TRUE
)

Arguments

chr
In the format 1:22 or X or Y.

plot_gene
Whether to plot positions of genes/transcripts on x-axis. Defaults to `FALSE`.

plot_exon
Whether to plot positions of exons on x-axis. Defaults to `FALSE`.

transcript_types
Genes/transcripts to be plotted. `"NM"` stands for protein-coding and `"NR"` stands for non-protein coding transcripts. By default, both types of transcripts are included.

title
Plot title. By default, omitted.

x_label
X-axis label. By default, omitted.

y_label
Y-axis label. By default, omitted.

text_cex
Numeric character expansion factor, used for all displayed text. Defaults to `1`.

beta_min
Minimum beta value to be displayed. Has to be in `[0,1>` range. Defaults to `0`.

beta_max
Maximum beta value to be displayed. Has to be in `<0,1]` range. Defaults to `1`.

error_size
Number of standard deviations which are to be plotted below and above the mean. Defaults to `1`.

plot_outside
Whether error plot should be plotted outside of the plotting area. Plotting area is determined with `beta_min` and `beta_max`. Defaults to `TRUE`.

Details

Plots error bars and confidence intervals for beta values of given CpGs. If list of CpGs is given, they are plotted equidistantly. If a range of coordinates and a chromosome are given, then all CpGs inside the range are plotted and positioned on x-axis based on their chromosomal coordinate. If the latter case, positions of unmeasured CpGs inside the region can be added to the plot.